+++ src/app/dashboard/dashboard.module.ts
@@ -7,21 +7,21 @@ import { DataService } from './services/';
 import { FileService } from './services/file.service';
 import { MonthlyDataComponent } from './monthly-data/monthly-data.component';
 import { UnitPipe } from './pipes/unit.pipe';
-import { PosnegStatisticDirective } from './directives/posneg-statistic.directive';
 import { StatisticItemComponent } from './components/statistic-item/statistic-item.component';
 import { MonthCitiesBarChartComponent } from './charts/month-cities-bar-chart/month-cities-bar-chart.component';
 import { DataVMService } from './services/data-vm.service';
 import { SettingsComponent } from './components/settings/settings.component';
+import { WithValuePipe } from './pipes/with-value.pipe';
 
 @NgModule({
   declarations: [
     DashboardIndexComponent,
     MonthlyDataComponent,
     UnitPipe,
-    PosnegStatisticDirective,
     StatisticItemComponent,
     MonthCitiesBarChartComponent,
     SettingsComponent,
+    WithValuePipe,
   ],
   imports: [
     SharedModule,