+++ src/app/dashboard/pipes/with-value.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { WithValuePipe } from './with-value.pipe';
+
+describe('WithValuePipe', () => {
+  it('create an instance', () => {
+    const pipe = new WithValuePipe();
+    expect(pipe).toBeTruthy();
+  });
+});